LayoutTests/http/tests/security/contentSecurityPolicy/1.1/securitypolicy-allowstylefrom.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
  <body>
    <script src="../resources/securitypolicy-tests-base.js"></script>
    <script>
      if (document.SecurityPolicy.allowsStyleFrom('http://example.com/'))
          log('PASS style is allowed when no policy exists.');
      else
          log('FAIL style is not allowed when no policy exists.');

      injectPolicy("style-src http://notexample.com;");

      if (!document.SecurityPolicy.allowsStyleFrom('http://example.com/'))
          log('PASS style is not allowed when policy exists.');
      else
          log('FAIL style is allowed when policy exists.');
    </script>
  </body>
</html>
