LayoutTests/http/tests/security/contentSecurityPolicy/1.1/securitypolicy-allowstylefrom-expected.txt

PASS style is allowed when no policy exists.
PASS style is not allowed when policy exists.
